test(execucao-ui): separa testes de cores e facade
