+++ web/execucao/tests/test_ui_execucao_ui_facade_templatetags.py
@@ -1,0 +1,8 @@
+from types import SimpleNamespace
+
+from web.execucao.templatetags.execucao_ui import projeto_color_bar
+
+
+def test_execucao_ui_reexports_projeto_color_bar():
+    projeto = SimpleNamespace(cor_slug="blue")
+    assert projeto_color_bar(projeto) == "bg-blue-500"
